+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Updated Exceptional call signature
+
 ## v0.18.17 (2024-08-21)
 * Converted consts to PascalCase
 * Made renderer consts protected